--- a/cpp/src/game/src/game/PlacementMove.cpp
+++ b/cpp/src/game/src/game/PlacementMove.cpp
@@ -1,20 +1,10 @@
 #include "game/PlacementMove.hpp"
 
 namespace Alphalcazar::Game {
-	PlacementMove::PlacementMove()
-		: Coordinates{ Coordinates::Invalid() }
-		, PieceType{ 0 }
-	{}
-
-	PlacementMove::PlacementMove(const Game::Coordinates& coordinates, Game::PieceType pieceType)
-		: Coordinates{ coordinates }
-		, PieceType{ pieceType }
-	{}
-
-	PlacementMove::~PlacementMove() {}
-
 	bool PlacementMove::Valid() const {
-		return Coordinates.Valid() && PieceType != 0;
+		bool coordinatesValid = Coordinates.Valid();
+		bool pieceTypeValid = PieceType != 0;
+		return coordinatesValid && pieceTypeValid;
 	}
 
 	bool PlacementMove::operator==(const PlacementMove& other) const {