Also removes the user-defined destructor and removes a branch in the `PlacementMove::Valid()` function. Does not significantly affect performance on its own, but becomes more relevant together with https://github.com/glecko/alphalcazar/pull/34.

Before:
```
[2022-09-04 11:20:41.717] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-09-04 11:20:41.720] [info] Game at depth 1 took 1ms ended with result 3
[2022-09-04 11:20:41.748] [info] First move at depth 2 took 27ms and calculated a score of -34
[2022-09-04 11:20:41.842] [info] Game at depth 2 took 93ms ended with result 3
[2022-09-04 11:20:42.243] [info] First move at depth 3 took 400ms and calculated a score of 35
[2022-09-04 11:20:50.458] [info] Game at depth 3 took 8213ms ended with result 2
[2022-09-04 11:22:23.987] [info] First move at depth 4 took 93528ms and calculated a score of -64
```

After:
```
[2022-09-04 11:18:07.201] [info] First move at depth 1 took 1ms and calculated a score of 0
[2022-09-04 11:18:07.205] [info] Game at depth 1 took 1ms ended with result 3
[2022-09-04 11:18:07.231] [info] First move at depth 2 took 25ms and calculated a score of -34
[2022-09-04 11:18:07.327] [info] Game at depth 2 took 95ms ended with result 3
[2022-09-04 11:18:07.734] [info] First move at depth 3 took 405ms and calculated a score of 35
[2022-09-04 11:18:15.998] [info] Game at depth 3 took 8264ms ended with result 2
[2022-09-04 11:19:52.160] [info] First move at depth 4 took 96161ms and calculated a score of -64
```